slides: How The Rust Compiler Will Assist You In Catching bugs
Rendered slides of my talk "How The Rust Compiler Will Assist You In
Catching Bugs". If you have a static web server, like the one that
Github Pages uses, you can serve these slides as a web page.

I added the notes of the presentation in the notes.md file because there
is no way with Sli.dev to render the slides with the presenter notes.